+++ coldfront/config/local_settings.py.sample
@@ -383,11 +383,14 @@ TOKEN_EXPIRATION_HOURS = 24
 BILLING_VALIDATOR_BACKEND = 'coldfront.core.billing.utils.validation.backends.dummy.DummyValidatorBackend'
 
 #------------------------------------------------------------------------------
-# Renewal Survey Validation settings
+# Renewal Survey Backend settings
 #------------------------------------------------------------------------------
 
 # The class to use for validating whether renewal survey has been completed.
-GOOGLE_RENEWAL_SURVEY_BACKEND = 'coldfront.core.project.utils_.google_renewal_survey.backends.dummy.DummyGoogleRenewalSurveyBackend'
+RENEWAL_SURVEY_BACKEND = 'coldfront.core.project.utils_.renewal_survey_backend.backends.permissive.PermissiveRenewalSurveyBackend'
+
+# Path to the renewal survey data file.
+RENEWAL_SURVEY_DATA_PATH = ''
 
 #------------------------------------------------------------------------------
 # Multiple Email Address settings
